partials, phoenix style
Another super lame, easy commit. TBH, I'm writing a lot
code in privates repos at the moment. I did think there
would be more to this commit than there was, which is why
the last commit was shorter as well. However, there are
several repl examples in the book which I worked through.