p. 73 registration changeset
The book has some accumulated errors with how changesets
are defined. This noted on the web, but the recommended
fixes don't work either, Phoenix has moved past them as
well.

This sort of thing isn't bad, per se. Fixing it will require
doing some open-ended digging, which means for now this
project is probably not suitable for daily incremental
progress. Daily increments need to be 15-30 minutes in
length, no longer.

This kind of blows as I'd committed to working through the
book on a regular schedule. I'll have to see about either
taking time from something else to fix it, or try to work
in the fixes after the usual morning commit.